если с какого-то перепуга codepage сделан uintptr_t, надо делать это везде

+++ far/fileedit.cpp
@@ -406,7 +406,7 @@ FileEditor::~FileEditor()
 
 void FileEditor::Init(
     const string& Name,
-    UINT codepage,
+    uintptr_t codepage,
     const string* Title,
     DWORD InitFlags,
     int StartLine,
